+++ components/Layout/Layout.jsx
@@ -1,6 +1,6 @@
 import Footer from "./Footer";
 import NavBar from "./NavBar";
-import styles from "../styles/Home.module.css";
+import styles from "../../styles/Layout.module.css";
 
 export default function Layout({ children }) {
   return (
