+++ components/Layout/Title.jsx
@@ -1,0 +1,12 @@
+import styles from "../../styles/Title.module.css";
+
+export default function Title() {
+  return (
+    <div>
+      <h1 className={styles.title}>Learn To Code</h1>
+      <p className={styles.description}>
+        Someone sent you this because of one of the following reasons:
+      </p>
+    </div>
+  );
+}
