--- components/Title.jsx
@@ -1,9 +1,0 @@
-import styles from "../styles/Home.module.css";
-
-export default function Title() {
-  return (
-    <div>
-      <h1 className={styles.title}>Learn To Code</h1>
-    </div>
-  );
-}
