+++ pages/about.jsx
@@ -1,8 +1,9 @@
-import Card from "../components/Card";
-import ExternalLink from "../components/ExternalLink";
+import Card from "../components/Card/Card";
+import ExternalLink from "../components/Utils/ExternalLink";
 import Image from "next/image";
-import PageHead from "../components/PageHead";
-import styles from "../styles/Home.module.css";
+import PageHead from "../components/Utils/PageHead";
+import styles from "../styles/Card.module.css";
+import utilStyles from "../styles/Utils.module.css";
 
 export default function AboutPage() {
   const profileSize = 144;
@@ -17,7 +18,7 @@ export default function AboutPage() {
               src="/profiles/retro-pfp.png"
               width={profileSize}
               height={profileSize}
-              className={styles.imageRound}
+              className={utilStyles.imageRound}
             />
             <p>
               Senior Moderator and Web Developer for Plexi Development.
@@ -31,7 +32,7 @@ export default function AboutPage() {
               src="/profiles/krysk-pfp.png"
               width={profileSize}
               height={profileSize}
-              className={styles.imageRound}
+              className={utilStyles.imageRound}
             />
             <p>
               Proficient User in Plexi Development. Apsiring Web/Software
