+++ styles/Footer.module.css
@@ -1,0 +1,22 @@
+.footer {
+  width: 100%;
+  height: 100px;
+  border-top: 1px solid #eaeaea;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.footer img {
+  margin-left: 0.5rem;
+}
+
+.footer a {
+  justify-content: center;
+  align-items: center;
+  color: var(--accent);
+}
+
+.footer .discordId {
+  font-size: 12px;
+}
