--- styles/Home.module.css
@@ -1,152 +1,0 @@
-.container {
-  min-height: 100vh;
-  padding: 0 0.5rem;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.main {
-  padding: 5rem 0;
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.footer {
-  width: 100%;
-  height: 100px;
-  border-top: 1px solid #eaeaea;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.footer img {
-  margin-left: 0.5rem;
-}
-
-.footer a {
-  justify-content: center;
-  align-items: center;
-  color: var(--accent);
-}
-
-.footer .id {
-  font-size: 12px;
-}
-
-.title a {
-  color: var(--accent);
-  text-decoration: none;
-}
-
-.title a:hover,
-.title a:focus,
-.title a:active {
-  text-decoration: underline;
-}
-
-.title {
-  margin: 0;
-  line-height: 1.15;
-  font-size: 4rem;
-}
-
-.title,
-.description {
-  text-align: center;
-}
-
-.description {
-  line-height: 1.5;
-  font-size: 1.5rem;
-}
-
-.code {
-  background: #fafafa;
-  border-radius: 10px;
-  padding: 20px;
-  font-size: 1.25rem;
-  font-family: Menlo, Monaco, Lucida Console, Liberation Mono, DejaVu Sans Mono,
-    Bitstream Vera Sans Mono, Courier New, monospace;
-}
-
-.grid {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-wrap: wrap;
-  max-width: 800px;
-  margin-top: 3rem;
-}
-
-.card {
-  margin: 1rem;
-  flex-basis: 45%;
-  padding: 1.5rem;
-  text-align: left;
-  color: inherit;
-  text-decoration: none;
-  border: 1px solid #eaeaea;
-  border-radius: 10px;
-  transition: color 0.15s ease, border-color 0.15s ease;
-}
-
-.card:hover,
-.card:focus,
-.card:active {
-  border-color: var(--accent);
-}
-
-.card h3 {
-  margin: 0 0 1rem 0;
-  font-size: 1.5rem;
-}
-
-.card p,
-.card ul li {
-  margin: 0;
-  font-size: 1.25rem;
-  line-height: 1.5;
-}
-
-.card ul {
-  list-style-type: none;
-}
-
-.card ul li:hover {
-  color: var(--accent);
-}
-
-.logo {
-  height: 1em;
-}
-
-.arrow {
-  vertical-align: middle;
-  padding-left: 10px;
-  transition: transform 0.2s ease-in-out;
-}
-
-.learnLink:hover .arrow {
-  transform: translate3d(25px, 0px, 0px);
-}
-
-.imageRound {
-  border-radius: 50%;
-}
-
-@media (max-width: 600px) {
-  .grid {
-    width: 100%;
-    flex-direction: column;
-  }
-
-  .learnLink .arrow {
-    display: none;
-  }
-}
