+++ styles/Title.module.css
@@ -1,0 +1,23 @@
+.title a {
+  color: var(--accent);
+}
+
+.title a:hover:focus:active {
+  text-decoration: underline;
+}
+
+.title {
+  margin: 0;
+  line-height: 1.15;
+  font-size: 4rem;
+}
+
+.title,
+.description {
+  text-align: center;
+}
+
+.description {
+  line-height: 1.5;
+  font-size: 1.5rem;
+}
